Rename "metronome" to "decode_metronome".
In preparation for experimentally supporting different types of
metronomes and metronome use cases we'd like to rename for clarity.

This is the first step, which introduces the new name and prefers it if
it is set, but keeps the old name for backwards compat reasons.

Once Chromium has migrated to the new name, we can delete the old name.

Bug: webrtc:15704
Change-Id: I23077bf2415ebb2b2338320c9a14e3bd17d3abb6