Make Codec::Matches also consider packetization
If it's not considered it can lead to payload IDs erroneously being
reused if the SDP is munged, see https://crbug.com/webrtc/15473#c10.

It's re-applicaiton of
https://webrtc-review.googlesource.com/c/src/+/326522. The
difference from that CL is in FindMatchingCodec that is made
to ignore packetization under some conditions. The general
rule is that FindMatchingCodec ignores packetization in
most contexts except when computing the set of used payloads.

Bug: webrtc:15473
Change-Id: Ice2b0f8d2c3aa0103ef69f8a5bfe27ba199ae970