dcsctp: Calculate RTO with webrtc::TimeDelta
This allows sub-millisecond RTOs (given that the socket is configured to
allow that low timeouts). But even without it, it allows higher
precision of the srtt() variable that is exported in metrics (although
not yet with that precision, but that's a fast follow).

This reverts to using the algorithm specified in RFC4960 - and not the
optimization to avoid floating point math - just for improved clarity.
All tests are unchanged, although slightly modified as it doesn't make
sense to verify exact microsecond precision.

Bug: webrtc:15593
Change-Id: Id30be4851ad35b42266f98cdc63e1a514ad0c8a7