p2p: Do not send Server Name Indication (SNI) if the server name is not set
Currently, SNI is sent even if the server name is not set. This is not
correct, as the server name is used to select the certificate to use
for the connection. For example, if TURNS server is given with an
IP address, the SNI should not be sent.

This patch fixes the issue by not sending SNI if the server name is not
set.

Bug: None
Change-Id: Ifd64ced02abeb35eec59a3dac2d76aad87299b46