[Stats] Move metric names to Attribute, constructed via AttributeInit.
As of this CL, Attribute no longer implements RTCStatsMemberInterface
and a member no longer owns knowing its own name. The attribute knows
the name because we pass it down at construction time.

To achieve this, the WEBRTC_RTCSTATS_IMPL() macro is updated to take
AttributeInits instead of raw member pointers, i.e. a (name, ptr) pair.

We now only use the default constructor of RTCStatsMember<T> which
behaves similar to a default constructed absl::optional<T> in that it
has no value by default, and now it has no name either. So we are very
close to being able to delete RTCStatsMember<T> in favor of
absl::optional<T>...

Bug: webrtc:15164
Change-Id: I560c0134bae1c2d7218426a1576425ecc1b677a7