[Stats] Move metric names to Attribute, constructed via AttributeInit.
As of this CL, Attribute no longer implements RTCStatsMemberInterface
and a member no longer owns knowing its own name. The attribute knows
the name because we pass it down at construction time.

To achieve this, the WEBRTC_RTCSTATS_IMPL() macro is updated to take
AttributeInits instead of raw member pointers, i.e. a (name, ptr) pair.

By now constructing RTCStatsMember<T> without a name parameter, it does
the same thing as the absl::optional<T> constructor. We should be able
to delete RTCStatsMember<T> as a follow-up CL.

Bug: webrtc:15164
Change-Id: I560c0134bae1c2d7218426a1576425ecc1b677a7