Fix Chromium roll failures due to classpath format changes in jni_zero
This commit addresses an issue resulting from a change [1] in the
jni_zero project, where the format of classpath entries changed
(from using slashes 'org/webrtc/PeerConnectionFactory'
to dots 'org.webrtc.PeerConnectionFactory'). These changes led to
failures in the Chromium rolls in WebRTC, as the Class loader in JNI
was not designed to handle class names with dots.

This CL fixes this issue by changing webrtc::GetClass to convert class
paths to what JNI expects.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/5234469

Bug: chromium:1377351
Change-Id: I2f243bb4ed04136f86510fcd5472e9bfc2d4ba85