Fix Android camera error of obtaining FramerateRange closest to the set frame rate.
Bug: None
Change-Id: Id209f8ea54560d650ddcc5577f68fcc4e0f4b634
The Android camera does not work well in finding the closest desired frame rate. It does not correctly calculate the weight for comparing frame rates. For example, if the desired frame rate is 15 frames per second, and the phone has two FramerateRanges: [12, 15] and [15, 15], the previous logic would match only [12, 15]. This commit fixes this issue.