[iOS] Make RTCRtpTransceiver.setCodecPreferences failable and accept nil
Web Spec and C++ version of setCodecPreferences are failable, as they return
an RTCError (in C++) or throw an InvalidModificationError (in Web Spec).
However, current Objective-C version of setCodecPreferences is not failable,
so callers cannot know if the operation succeeded or not.
Also, the current Objective-C version does not accept nil, which is not
spec-compliant. (Web Spec says if codecs is an empty list, set
transceiver.PreferredCodecs to codecs and abort these steps.)

Bug: webrtc:42226156
Change-Id: Ib90f3e5b45fc959eeb92f623cf50efcb458a7478